+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# scorecardpy 0.1.6.5
+# scorecardpy 0.1.7
 
 * pdo in scorecard function now suports negative value. If pdo is positive, the larger score means the lower probability to be positive sample. If pdo is negative, the larger score means the higher probability to be positive sample.
 * fixed a bug in woebin function using chimerge method, which is caused by initial breaks have out-range values.
@@ -9,6 +9,7 @@
 * If the argument return_rm_reason is TRUE in var_filter function, the info_value, missing_rate and identical_rate are provided.
 * fixed a bug in woebin_adj that breaks do not update when chooseing next. 
 * fixed a bug is woebin function cant modify positive values
+* check duplicated index in input dataframe
 
 # scorecardpy 0.1.6